Move json code to document package
