--- types/json.go
@@ -1,71 +1,0 @@
-package types
-
-import (
-	"bytes"
-	"strconv"
-)
-
-type JsonArray struct {
-	Array
-}
-
-func (j JsonArray) MarshalJSON() ([]byte, error) {
-	var buf bytes.Buffer
-
-	buf.WriteByte('[')
-	var notFirst bool
-	err := j.Array.Iterate(func(i int, v Value) error {
-		if notFirst {
-			buf.WriteString(", ")
-		}
-		notFirst = true
-
-		data, err := v.MarshalJSON()
-		if err != nil {
-			return err
-		}
-
-		_, err = buf.Write(data)
-		return err
-	})
-	if err != nil {
-		return nil, err
-	}
-	buf.WriteByte(']')
-	return buf.Bytes(), nil
-}
-
-type JsonDocument struct {
-	Document
-}
-
-func (j JsonDocument) MarshalJSON() ([]byte, error) {
-	var buf bytes.Buffer
-
-	buf.WriteByte('{')
-
-	var notFirst bool
-	err := j.Document.Iterate(func(f string, v Value) error {
-		if notFirst {
-			buf.WriteString(", ")
-		}
-		notFirst = true
-
-		buf.WriteString(strconv.Quote(f))
-		buf.WriteString(": ")
-
-		data, err := v.MarshalJSON()
-		if err != nil {
-			return err
-		}
-		_, err = buf.Write(data)
-		return err
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	buf.WriteByte('}')
-
-	return buf.Bytes(), nil
-}
